Update add-edit-project.component.ts
Removed the change in this commit until we have have a better look at CKEditor. The EventInfo type doesn't seem to be included within our custom CKEditor instance (despite CKEditor using it), and installing additional packages could interfere with our use of CK. In the future we'll need to look in more depth, but out of scope for this ticket.